feat: Add react-icons package and update package.json
- Added the "react-icons" package to the frontend's package.json file.
- This package provides a collection of popular icons for React applications.
- The package version is updated in the package.json file.

Refactor viewsFormatter.js

- Removed a console.log statement from the viewsFormatter.js file.

feat: Add isLiked field to like.model.js

- Added a new field "isLiked" to the like.model.js file in the backend.
- This field is of type Boolean and has a default value of true.

Refactor store.js

- Renamed the "isLogin" reducer to "auth" in the store.js file.

Update subscriptions.model.js

- Changed the type of the "channel" field in the subscriptions.model.js file to Schema.Types.ObjectId.

Update isLoginSlice.js

- Renamed the "isLogin" slice to "auth" in the isLoginSlice.js file.

Update cloudinary.js

- Commented out the fs.unlinkSync() line and added a call to the safeUnlink() function in the catch block of the uploadOnCloudinary() function in the cloudinary.js file.

Update ChannelLayout.jsx

- Added a new line after the Outlet component in the ChannelLayout.jsx file.

Add sharedGetCurrentUser.js

- Added a new file sharedGetCurrentUser.js to the frontend's Utils/sharedQuaries directory.

Update VideoPage.jsx

- Updated the useSelector() function in the VideoPage.jsx file to use the "auth" slice instead of the "isLogin" slice.

Update subscription.controllers.js

- Added a console.log statement to the toggleSubscription function in the subscription.controllers.js file.

Update like.routes.js

- Added a new route "/getDisLikes/:id" to the like.routes.js file.

Update Layout.jsx

- Added a useEffect() hook to the Layout.jsx file to log a message when the page is refreshed.

Update video.routes.js

- Moved the getAllVideos route above the verifyJWT middleware in the video.routes.js file.

Update Login.jsx

- Updated the useSelector() function in the Login.jsx file to use the "auth" slice instead of the "isLogin" slice.

Update user.routes.js

- Added new routes "/likedVideos" and "/myVideos" to the user.routes.js file.

Update package.json

- Updated the package-lock.json file to reflect the addition of the "react-icons" package.

Update video.controllers.js

- Added validation for the title and description fields in the publishAVideo function in the video.controllers.js file.

Update main.jsx

- Added a new route "/videoList/:v" to the main.jsx file.

+++ backend/src/controllers/like.controllers.js
@@ -6,23 +6,45 @@ import { asyncHandler } from "../utils/asyncHandler.js";
 
 const toggleVideoLike = asyncHandler(async (req, res) => {
     const { videoId } = req.params;
+    let { isLiked } = req.body;
     //TODO: toggle like on video
+
+    console.log("entered toggle like" , isLiked);
+    
     if (!videoId) {
         throw new ApiError(400, "Video id is missing");
     }
+    // isLiked = isLiked === "true" ? true : false;
 
+    console.log("isLiked: ", isLiked);
+    
     const likePrev = await Like.find({ video: videoId, likedBy: req.user._id });
 
     let flag = false;
-    let like;
-    if (likePrev.length > 0) {
+    let like = {};
+    if (likePrev?.length > 0 && !(likePrev[0].isLiked ^ isLiked)) {
         flag = true;
-        like = await Like.findByIdAndDelete(likePrev[0]._id);
+
+        await Like.findByIdAndDelete(likePrev[0]._id);
+    } else if (likePrev?.length > 0 && likePrev[0].isLiked ^ isLiked) {
+        flag = false;
+
+        like = await Like.findByIdAndUpdate(
+            likePrev[0]._id,
+            {
+                isLiked,
+            },
+            {
+                new: true,
+            }
+        );
     } else {
         flag = false;
+
         like = await Like.create({
             video: videoId,
             likedBy: req.user._id,
+            isLiked,
         });
     }
 
@@ -39,7 +61,9 @@ const toggleVideoLike = asyncHandler(async (req, res) => {
         new ApiResponse(
             200,
             like,
-            flag ? "Video disliked successfully" : "Video liked successfully"
+            flag
+                ? "Video deleted successfully"
+                : "Video liked / disliked successfully"
         )
     );
 });
@@ -159,31 +183,75 @@ const getLikeCount = asyncHandler(async (req, res) => {
     }
     const likeCount = await Like.aggregate([
         {
-            $match : {
-                $or : [
-                    {video : new mongoose.Types.ObjectId(id)},
-                    {comment : new mongoose.Types.ObjectId(id)},
-                    {tweet : new mongoose.Types.ObjectId(id)}
-                ]
-            }
+            $match: {
+                $or: [
+                    { video: new mongoose.Types.ObjectId(id) },
+                    { comment: new mongoose.Types.ObjectId(id) },
+                    { tweet: new mongoose.Types.ObjectId(id) },
+                ],
+                isLiked: true,
+            },
         },
         {
-            $count : "totalLikes"
+            $count: "likeCount",
         },
         {
-            $project : {
-                totalLikes : 1
-            }
-        }
+            $project: {
+                likeCount: 1,
+            },
+        },
     ]);
     console.log(likeCount);
-    
-    
 
     res.status(200).json(
-        new ApiResponse(200, likeCount[0] || {likeCount : 0}, "Like count fetched successfully")
+        new ApiResponse(
+            200,
+            likeCount[0] || { likeCount: 0 },
+            "Like count fetched successfully"
+        )
     );
+});
+const getDisLikeCount = asyncHandler(async (req, res) => {
+    const { id } = req.params;
 
-}
-)
-export { toggleCommentLike, toggleTweetLike, toggleVideoLike, getLikedVideos , getLikeCount };
+    if (!id) {
+        throw new ApiError(400, "Id is missing");
+    }
+    const disLikeCount = await Like.aggregate([
+        {
+            $match: {
+                $or: [
+                    { video: new mongoose.Types.ObjectId(id) },
+                    { comment: new mongoose.Types.ObjectId(id) },
+                    { tweet: new mongoose.Types.ObjectId(id) },
+                ],
+                isLiked: false,
+            },
+        },
+        {
+            $count: "disLikeCount",
+        },
+        {
+            $project: {
+                disLikeCount: 1,
+            },
+        },
+    ]);
+    console.log(disLikeCount);
+
+    res.status(200).json(
+        new ApiResponse(
+            200,
+            disLikeCount[0] || { disLikeCount: 0 },
+            "DisLikeLike count fetched successfully"
+        )
+    );
+});
+export {
+    toggleCommentLike,
+    toggleTweetLike,
+    toggleVideoLike,
+    getLikedVideos,
+    getLikeCount,
+    getDisLikeCount,
+};
